refactor pages to server components
